# Improve performance of style class selector subscriptions