## What does the pull request do?

Previously, when a selector such as `Button.foo` was instanced on a control, the style instance attached a listener to the style class selector by subscribing to `INotifyCollectionChanged.CollectionChanged`. However it has been shown in profiling (such as in #8389) that adding these event handlers is a performance bottleneck. 

This PR adds an internal listener API to `Classes` which simply stores a list of listeners in a list. For the repro in #8389, on my machine this improved time to detach the control from around 1.2s to 0.273s.

This PR also adds a benchmark to test this behavior. Here are the numbers:

`master`

|       Method |       Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------- |-----------:|---------:|---------:|-------:|-------:|----------:|
|        Apply |   920.5 ns |  3.20 ns |  2.68 ns | 0.1526 | 0.0049 |     642 B |
| Apply_Toggle | 1,639.2 ns | 12.04 ns | 11.27 ns | 0.1929 |      - |     810 B |
| Apply_Detach | 1,895.0 ns |  5.09 ns |  4.76 ns | 0.2588 | 0.0024 |   1,090 B |

This PR:

|       Method |       Mean |   Error |  StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------- |-----------:|--------:|--------:|-------:|-------:|----------:|
|        Apply |   808.2 ns | 5.18 ns | 4.33 ns | 0.1208 |      - |     507 B |
| Apply_Toggle | 1,528.9 ns | 5.81 ns | 4.85 ns | 0.1587 |      - |     671 B |
| Apply_Detach | 1,258.6 ns | 7.19 ns | 6.01 ns | 0.1392 | 0.0024 |     588 B |

I also tried using a `HashSet` instead of a list, given that handlers can be removed from anywhere in the list, but this had worse performance both in the repro to #8389 and the benchmark:

`HashSet` implementation:

`hashset`

|       Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------- |-----------:|---------:|---------:|-------:|----------:|
|        Apply |   877.6 ns | 13.66 ns | 15.18 ns | 0.1318 |     552 B |
| Apply_Toggle | 1,527.2 ns |  8.86 ns |  7.86 ns | 0.1709 |     716 B |
| Apply_Detach | 1,267.5 ns |  7.72 ns |  6.84 ns | 0.1489 |     633 B |

## Fixed issues

Fixes #8389