@avaloniaui-team 2022-07-03
You can test this PR using the following package version. `0.10.999-cibuild0021528-beta`. (feed url: https://nuget.avaloniaui.net/repository/avalonia-all/index.json) [PRBUILDID]