@timunie 2022-07-07
@pr8x maybe this should be a feature request to not forget it? Sounds good to me